refactor: Move and rename files to provide better structure and avoid known issues with Styles folder
