+++ src/DevTKSS.Uno.Samples.MvuxGallery/Models/GalleryImages/GalleryImage.cs
