+++ src/DevTKSS.Uno.Samples.MvuxGallery/Services/ICodeSampleService.cs
@@ -1,5 +1,5 @@
 
-namespace DevTKSS.Uno.Samples.MvuxGallery.Models.CodeSamples;
+namespace DevTKSS.Uno.Samples.MvuxGallery.Services;
 
 public interface ICodeSampleService
 {
@@ -10,7 +10,7 @@ public interface ICodeSampleService
     /// <param name="sampleID">The identifier of the sample.</param>
     /// <param name="ct">A cancellation token for the operation.</param>
     /// <returns>The content of the code sample.</returns>
-    public ValueTask<string> GetCodeSampleAsync(string sampleID,CancellationToken ct = default);
+    public ValueTask<string> GetCodeSampleAsync(string sampleID, CancellationToken ct = default);
 
     /// <summary>
     /// Get a static collection of values for code sample options asynchronously.
